+++ rust/sedona-expr/src/function_set.rs
@@ -146,11 +146,11 @@ mod tests {
     use datafusion_common::{not_impl_err, scalar::ScalarValue};
 
     use datafusion_expr::{Accumulator, ColumnarValue, Volatility};
-    use sedona_schema::datatypes::SedonaType;
+    use sedona_schema::{datatypes::SedonaType, matchers::ArgMatcher};
 
     use crate::{
         aggregate_udf::SedonaAccumulator,
-        scalar_udf::{ArgMatcher, SimpleSedonaScalarKernel},
+        scalar_udf::{SimpleSedonaScalarKernel},
     };
 
     use super::*;
